create new write article page
